Moved IS-12 command support into IS12Utils
